Renamed gqlApollo to googleCloudApollo
