+++ packages/apollo-server-cloud-function/src/googleCloudApollo.ts
